Add APICall.swift for API management and endpoint definitions
