+++ GimOn/NetworkSource/NetworkService.swift
@@ -7,17 +7,6 @@
 
 import Foundation
 
-struct API {
-    static let baseUrl = "https://api.rawg.io/api/games"
-    
-    static let apiKey: String = {
-        guard let apiKey = Bundle.main.object(forInfoDictionaryKey: "API_KEY") as? String else {
-            fatalError("API_KEY not found in Info.plist")
-        }
-        return apiKey
-    }()
-}
-
 class NetworkService {
     
     let pageSize = "35"
